better autotrimming implemented in covest
