+++ bin/hist_analyze.py
@@ -3,14 +3,7 @@ import argparse
 
 from covest import config
 from covest.data import load_histogram
-from covest.histogram import auto_sample_hist, get_trim
-
-
-def remove_noise(hist):
-    total = sum(hist.values())
-    hist_denoised = {k: v for k, v in hist.items() if v/total > config.NOISE_THRESHOLD}
-    tail = sum(v for v in hist.values() if v/total <= config.NOISE_THRESHOLD)
-    return hist_denoised, tail
+from covest.histogram import auto_sample_hist, get_trim, remove_noise
 
 
 def main(args):
@@ -19,7 +12,7 @@ def main(args):
     hist_orig, tail = remove_noise(hist_orig)
     hist, sample_factor = auto_sample_hist(hist_orig, k, r)
     print('Suggested sf:', sample_factor)
-    print('Suggested trim:', max(get_trim(hist, 6), min(50, max(hist))))
+    print('Suggested trim:', max(get_trim(hist), min(50, max(hist))))
 
 
 if __name__ == '__main__':
