+++ covest/config.py
@@ -15,7 +15,9 @@ DEFAULT_READ_LENGTH = 100
 DEFAULT_REPEAT_MODEL = 0
 DEFAULT_MIN_SINGLECOPY_RATIO = 0.3
 AUTO_SAMPLE_TARGET_COVERAGE = 12
+AUTO_TRIM_PRECISION = 6
 NOISE_THRESHOLD = 10**-6
+
 try:
     from multiprocessing import cpu_count
